+++ src/ts/beatmap/environment.ts
@@ -45,6 +45,8 @@ export type EnvironmentName =
     | 'BTSEnvironment'
     | 'KaleidoscopeEnvironment'
     | 'InterscopeEnvironment'
+    | 'SkrillexEnvironment'
+    | 'BillieEnvironment'
     | 'GlassDesertEnvironment';
 
 export enum EnvironmentRename {
@@ -68,6 +70,8 @@ export enum EnvironmentRename {
     BTSEnvironment = 'BTS',
     KaleidoscopeEnvironment = 'Kaleidoscope',
     InterscopeEnvironment = 'Interscope',
+    SkrillexEnvironment = 'Skrillex',
+    BillieEnvironment = 'Billie',
     GlassDesertEnvironment = 'Glass Desert',
 }
 
@@ -92,6 +96,8 @@ export enum EnvironmentColor {
     BTSEnvironment = 'BTS',
     KaleidoscopeEnvironment = 'Kaleidoscope',
     InterscopeEnvironment = 'Interscope',
+    SkrillexEnvironment = 'Skrillex',
+    BillieEnvironment = 'Billie Eilish',
     GlassDesertEnvironment = 'Glass Desert',
 }
 
@@ -512,6 +518,80 @@ export const colorScheme: EnvironmentScheme = {
             r: 0.588235,
             g: 0.298039,
             b: 0.721569,
+        },
+    },
+    Skrillex: {
+        _colorLeft: {
+            r: 0.69803923,
+            g: 0.14117648,
+            b: 0.36862746,
+        },
+        _colorRight: {
+            r: 0.32933334,
+            g: 0.32299998,
+            b: 0.38,
+        },
+        _envColorLeft: {
+            r: 0.80000001,
+            g: 0.28000003,
+            b: 0.58594489,
+        },
+        _envColorRight: {
+            r: 0.06525807,
+            g: 0.57800001,
+            b: 0.56867743,
+        },
+        _envColorLeftBoost: {
+            r: 0.81176478,
+            g: 0.30588236,
+            b: 0.30588236,
+        },
+        _envColorRightBoost: {
+            r: 0.27843139,
+            g: 0.80000001,
+            b: 0.44597632,
+        },
+        _obstacleColor: {
+            r: 0.15686275,
+            g: 0.60392159,
+            b: 0.60392159,
+        },
+    },
+    'Billie Eilish': {
+        _colorLeft: {
+            r: 0.80000001,
+            g: 0.64481932,
+            b: 0.43200001,
+        },
+        _colorRight: {
+            r: 0.54808509,
+            g: 0.61276591,
+            b: 0.63999999,
+        },
+        _envColorLeft: {
+            r: 0.81960785,
+            g: 0.442,
+            b: 0.184,
+        },
+        _envColorRight: {
+            r: 0.94117647,
+            g: 0.70677096,
+            b: 0.56470591,
+        },
+        _obstacleColor: {
+            r: 0.71325314,
+            g: 0.56140977,
+            b: 0.78301889,
+        },
+        _envColorLeftBoost: {
+            r: 0.80000001,
+            g: 0,
+            b: 0,
+        },
+        _envColorRightBoost: {
+            r: 0.55686277,
+            g: 0.7019608,
+            b: 0.77647066,
         },
     },
     'Glass Desert': {