+++ src/ts/tools/events/index.ts
@@ -1,1 +1,3 @@
 export { default as insufficientLight } from './insufficientLight';
+export { default as oldEventValue } from './oldEventValue';
+export { default as testEvent } from './testEvent';