+++ src/ts/tools/events/oldEventValue.ts
@@ -1,15 +1,14 @@
 import * as beatmap from '../../beatmap';
 import { round } from '../../utils';
 import { BeatmapSettings, Tool } from '../template';
-import * as swing from '../swing';
 
 const htmlContainer = document.createElement('div');
 const htmlInputCheck = document.createElement('input');
 const htmlLabelCheck = document.createElement('label');
 
-htmlLabelCheck.textContent = ' Hitbox inline';
-htmlLabelCheck.htmlFor = 'input__tools-hitbox-inline-check';
-htmlInputCheck.id = 'input__tools-hitbox-inline-check';
+htmlLabelCheck.textContent = ' Old value 4 event';
+htmlLabelCheck.htmlFor = 'input__tools-light-4';
+htmlInputCheck.id = 'input__tools-light-4';
 htmlInputCheck.className = 'input-toggle';
 htmlInputCheck.type = 'checkbox';
 htmlInputCheck.checked = true;
@@ -19,12 +18,12 @@ htmlContainer.appendChild(htmlInputCheck);
 htmlContainer.appendChild(htmlLabelCheck);
 
 const tool: Tool = {
-    name: 'Hitbox Inline',
+    name: 'Insufficient Light',
     description: 'Placeholder',
-    type: 'note',
+    type: 'event',
     order: {
-        input: 72,
-        output: 192,
+        input: 0,
+        output: 0,
     },
     input: {
         enabled: htmlInputCheck.checked,
@@ -42,42 +41,18 @@ function inputCheckHandler(this: HTMLInputElement) {
 }
 
 function check(mapSettings: BeatmapSettings, mapSet: beatmap.map.BeatmapSetData) {
-    const { _bpm: bpm, _njs: njs } = mapSettings;
-    const { _notes: notes } = mapSet._data;
+    const { _bpm: bpm, _audioDuration: duration } = mapSettings;
+    const { _events: events } = mapSet._data;
 
-    const lastNote: { [key: number]: beatmap.note.Note } = {};
-    const swingNoteArray: { [key: number]: beatmap.note.Note[] } = {
-        0: [],
-        1: [],
-        3: [],
-    };
-
-    const arr: beatmap.note.Note[] = [];
-    for (let i = 0, len = notes.length; i < len; i++) {
-        const note = notes[i];
-        if (beatmap.note.isNote(note) && lastNote[note._type]) {
-            if (
-                swing.next(note, lastNote[note._type], bpm, swingNoteArray[note._type])
-            ) {
-                swingNoteArray[note._type] = [];
+    const arr: beatmap.event.Event[] = [];
+    if (beatmap.version.compare(mapSet._data._version, 'difficulty') === 'old') {
+        for (let i = 0; i < events.length; i++) {
+            if (beatmap.event.isLightEvent(events[i]) && events[i]._value === 4) {
+                arr.push(events[i]);
             }
         }
-        for (const other of swingNoteArray[(note._type + 1) % 2]) {
-            // magic number 1.425 from saber length + good/bad hitbox
-            if (
-                !(
-                    njs.value >
-                    (1.425 * bpm.value) / (60 * (note._time - other._time))
-                ) &&
-                note._lineIndex === other._lineIndex &&
-                note._lineLayer === other._lineLayer
-            ) {
-                arr.push(note);
-                break;
-            }
-        }
-        lastNote[note._type] = note;
-        swingNoteArray[note._type].push(note);
+    } else {
+        return [];
     }
     return arr
         .map((n) => n._time)
@@ -94,7 +69,7 @@ function run(mapSettings: BeatmapSettings, mapSet?: beatmap.map.BeatmapSetData):
 
     if (result.length) {
         const htmlResult = document.createElement('div');
-        htmlResult.innerHTML = `<b>Hitbox inline [${result.length}]:</b> ${result
+        htmlResult.innerHTML = `<b>Event with value 4 [${result.length}]:</b> ${result
             .map((n) => round(mapSettings._bpm.adjustTime(n), 3))
             .join(', ')}`;
         tool.output.html = htmlResult;