+++ src/ts/ui/index.ts
@@ -1,7 +1,7 @@
 // ui is massive fuckin w
 export * as accordion from './accordion';
 export * as header from './header';
-export * as info from './info';
+export * as info from './information';
 export * as input from './input';
 export * as intro from './intro';
 export * as loading from './loading';